backend/pyproject.toml
[tool.poetry]
name = "college-enrolment-system"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
dj-rest-auth = {extras = ["with-social"], version = "^6.0.0"}
drf-yasg = "^1.21.7"
black = "^24.4.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
